+++ srcs/print_str.c
@@ -6,7 +6,7 @@
 /*   By: wchen <wchen@42studen>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/10 20:39:26 by wchen             #+#    #+#             */
-/*   Updated: 2022/10/11 22:40:26 by wchen            ###   ########.fr       */
+/*   Updated: 2022/10/12 23:46:15 by wchen            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,10 +19,10 @@ int	print_str(char *str)
 
 	if (!str)
 	{
-		ft_putstr_fd("(null)", (int)1);
+		ft_putstr_fd("(null)", 1);
 		return (6);
 	}
-	ft_putstr_fd(str, (int)1);
+	ft_putstr_fd(str, 1);
 	len_str = ft_strlen(str);
 	return (len_str);
 }
